cephadm: move CephContainer/similar to new container_types.py
Part of general cephadm split-up refactoring. I am not happy with the
name 'container_types' but none of the alternatives I could think
of were much better.
